refactor: split roadmap route into focused modules
The roadmap route was a single 906-line file mixing types, config,
168 lines of mock data, three full view implementations, and the page
shell. Split along the existing seams (no behavior change):

- -types.ts        RoadmapItem / RoadmapStatus / ViewMode / StatusConfig
- -config.ts       STATUS_CONFIG, STATUSES, LIST_ORDER, QUARTERS
- -mock-data.ts    MOCK_ITEMS (placeholder until backend-wired)
- -components/board-view.tsx     BoardCard + Draggable/Droppable + BoardView
- -components/list-view.tsx      ListItem + ListView
- -components/timeline-view.tsx  TimelineCard + TimelineView

index.tsx is now just the route + toolbar page (906 -> 99 lines). The
`-`-prefixed files are excluded from routing, so routeTree.gen.ts is
unchanged.

Verified: tsc clean, 119/119 tests pass, prettier clean.
