+++ src/routes/@{$org}/$project/roadmap/-mock-data.ts
@@ -1,0 +1,172 @@
+import type { RoadmapItem } from "./-types"
+
+// TODO: replace with real roadmap data once wired to the backend.
+export const MOCK_ITEMS: Array<RoadmapItem> = [
+  // Backlog
+  {
+    id: "1",
+    title: "Advanced search with smart filters",
+    status: "backlog",
+    tags: ["Feedback", "UX"],
+    upvotes: 47,
+    feedbackCount: 12,
+    githubIssues: 0,
+    quarter: "Q4 2025",
+  },
+  {
+    id: "2",
+    title: "Custom notification preferences",
+    status: "backlog",
+    tags: ["Settings"],
+    upvotes: 31,
+    feedbackCount: 8,
+    githubIssues: 0,
+    quarter: "Q4 2025",
+  },
+  {
+    id: "3",
+    title: "Data export (CSV & JSON)",
+    status: "backlog",
+    tags: ["Data", "Admin"],
+    upvotes: 89,
+    feedbackCount: 23,
+    githubIssues: 2,
+    quarter: "Q1 2026",
+  },
+  {
+    id: "4",
+    title: "Team @mentions in comments",
+    status: "backlog",
+    tags: ["Collaboration"],
+    upvotes: 22,
+    feedbackCount: 6,
+    githubIssues: 0,
+    quarter: "Q1 2026",
+  },
+
+  // Planned
+  {
+    id: "5",
+    title: "Roadmap public sharing & embed",
+    status: "planned",
+    tags: ["Roadmap", "Public"],
+    upvotes: 134,
+    feedbackCount: 31,
+    githubIssues: 3,
+    quarter: "Q3 2025",
+  },
+  {
+    id: "6",
+    title: "GitHub issue bidirectional sync",
+    status: "planned",
+    tags: ["GitHub", "Integration"],
+    upvotes: 203,
+    feedbackCount: 44,
+    githubIssues: 7,
+    quarter: "Q3 2025",
+  },
+  {
+    id: "7",
+    title: "Custom status labels per project",
+    status: "planned",
+    tags: ["Customization"],
+    upvotes: 56,
+    feedbackCount: 14,
+    githubIssues: 1,
+    quarter: "Q3 2025",
+  },
+  {
+    id: "8",
+    title: "Admin upvote limit configuration",
+    status: "planned",
+    tags: ["Admin"],
+    upvotes: 18,
+    feedbackCount: 5,
+    githubIssues: 0,
+    quarter: "Q4 2025",
+  },
+
+  // In Progress
+  {
+    id: "9",
+    title: "Feedback voting system v2",
+    status: "in-progress",
+    tags: ["Feedback", "Core"],
+    upvotes: 312,
+    feedbackCount: 67,
+    githubIssues: 11,
+    quarter: "Q3 2025",
+  },
+  {
+    id: "10",
+    title: "Integration marketplace",
+    status: "in-progress",
+    tags: ["Integrations"],
+    upvotes: 178,
+    feedbackCount: 39,
+    githubIssues: 5,
+    quarter: "Q3 2025",
+  },
+  {
+    id: "11",
+    title: "Rich text comment editor",
+    status: "in-progress",
+    tags: ["Comments", "UX"],
+    upvotes: 91,
+    feedbackCount: 21,
+    githubIssues: 2,
+    quarter: "Q3 2025",
+  },
+
+  // Released
+  {
+    id: "12",
+    title: "Webhook integrations",
+    status: "released",
+    tags: ["Integrations", "API"],
+    upvotes: 445,
+    feedbackCount: 89,
+    githubIssues: 14,
+    quarter: "Q2 2025",
+  },
+  {
+    id: "13",
+    title: "SSO & SAML support",
+    status: "released",
+    tags: ["Auth", "Enterprise"],
+    upvotes: 521,
+    feedbackCount: 103,
+    githubIssues: 8,
+    quarter: "Q2 2025",
+  },
+  {
+    id: "14",
+    title: "Comment emoji reactions",
+    status: "released",
+    tags: ["Comments", "UX"],
+    upvotes: 234,
+    feedbackCount: 52,
+    githubIssues: 3,
+    quarter: "Q1 2025",
+  },
+  {
+    id: "15",
+    title: "CSV data import",
+    status: "released",
+    tags: ["Data", "Admin"],
+    upvotes: 167,
+    feedbackCount: 38,
+    githubIssues: 6,
+    quarter: "Q1 2025",
+  },
+  {
+    id: "16",
+    title: "Email digest notifications",
+    status: "released",
+    tags: ["Notifications"],
+    upvotes: 389,
+    feedbackCount: 77,
+    githubIssues: 9,
+    quarter: "Q1 2025",
+  },
+]
